GenAI_Course/GenAI201_Project/requirements.txt
streamlit
langchain
transformers
chromadb
mysql-connector-python
python-dotenv
notebook